fix circular dependency for tests
