+++ packages/legend-application-query/src/stores/QueryBuilderFetchStructureState.ts
@@ -18,6 +18,7 @@ import { action, makeAutoObservable } from 'mobx';
 import type { QueryBuilderState } from './QueryBuilderState.js';
 import { QueryBuilderGraphFetchTreeState } from './QueryBuilderGraphFetchTreeState.js';
 import { QueryBuilderProjectionState } from './QueryBuilderProjectionState.js';
+import type { QueryBuilderAggregateOperator } from './QueryBuilderAggregationState.js';
 
 export enum FETCH_STRUCTURE_MODE {
   PROJECTION = 'PROJECTION',
@@ -30,7 +31,10 @@ export class QueryBuilderFetchStructureState {
   projectionState: QueryBuilderProjectionState;
   graphFetchTreeState: QueryBuilderGraphFetchTreeState;
 
-  constructor(queryBuilderState: QueryBuilderState) {
+  constructor(
+    queryBuilderState: QueryBuilderState,
+    operators: QueryBuilderAggregateOperator[],
+  ) {
     makeAutoObservable(this, {
       queryBuilderState: false,
       setFetchStructureMode: action,
@@ -39,7 +43,10 @@ export class QueryBuilderFetchStructureState {
     this.queryBuilderState = queryBuilderState;
     // TODO: we probably should modularize this a bit better
     // See https://github.com/finos/legend-studio/issues/731
-    this.projectionState = new QueryBuilderProjectionState(queryBuilderState);
+    this.projectionState = new QueryBuilderProjectionState(
+      queryBuilderState,
+      operators,
+    );
     this.graphFetchTreeState = new QueryBuilderGraphFetchTreeState(
       queryBuilderState,
     );
